requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
Pillow>=10.0.0
qwen-vl-utils 